requirements.txt
# RAG System - Puls-Events POC

# Vector search
faiss-cpu>=1.7.4

# LangChain ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-mistralai>=0.1.0
langchain-text-splitters>=0.2.0

# Embeddings
sentence-transformers>=2.7.0
langchain-huggingface>=0.1.0

# Mistral AI
mistralai>=1.0.0

# API
fastapi>=0.111.0
uvicorn>=0.30.0

# Data
pandas>=2.2.0
requests>=2.32.0

# Evaluation
ragas==0.2.15

# Utils
python-dotenv>=1.0.0
httpx>=0.27.0
pytest==9.0.2
pytest-cov>=5.0.0
